+++ cmd/owl/session/session.go
@@ -3,6 +3,7 @@ package session
 import (
 	"fmt"
 	"io/ioutil"
+	"os"
 
 	yaml "gopkg.in/yaml.v3"
 )
@@ -24,8 +25,12 @@ func NewSession(sourcefile string) *Session {
 // Restore session from file
 func (s *Session) Restore() (*Session, error) {
 	dat, err := ioutil.ReadFile(s.sourcefile)
+	if os.IsNotExist(err) {
+		err = s.Dump()
+	}
+
 	if err != nil {
-		return s, fmt.Errorf("invalid session file: %s", err)
+		return s, fmt.Errorf("error reading session file: %s", err)
 	}
 
 	err = yaml.Unmarshal(dat, s)